nak_controller/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tradepulse-nak"
version = "2.1.0"
description = "NaK Neuro-Energetic Controller for multi-strategy trading systems"
authors = [{name="TradePulse R&D"}]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "numpy>=1.24",
  "pyyaml>=6.0",
  "pydantic>=2.7",
]

[project.scripts]
nak-cv = "nak_controller.cli.run_cv:main"
nak-validate = "nak_controller.cli.run_validate:main"

[tool.setuptools]
packages = ["nak_controller"]
package-dir = { "nak_controller" = "." }
include-package-data = true

[tool.setuptools.package-data]
"nak_controller" = ["conf/*.yaml"]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["."]
